Inline startup message and remove key hints
